Sprite loading and assignment
The effect is a bit underwhelming, it just shows two squares (that are
more like rectangles anyway due to faulty camera setup). But it goes
through the whole machinery of loading sprites (someone needs to draw
something nicer than just colorful squares), entity creation and
addition of the sprites to them automatically.